General clean up and move most types to appropriate modules
Moved the tutorial material in `C.A` to `C.A.Tutorial`.

Renamed `Location` to `IniVariable` and used everywhere instead of having
section and variable split out in `Info` and `ExampleItem`.

Renamed `Example` to `Sample` in `C.A.Types` since the `Example` name is used so
heavily in the `C.A.Example` module.  The overlap was confusing.